Grid Plugin First Refactor (#397)
* tests passing but shouldnt be

* more int -> uint

* remove previous grid blocks

* some renaming

* remove teardown

* remove more static structs from plugin?

* renaming

* create json

* more changes

* address PR comments

* remove duplicate temp from erd_gui

* change cpp properties

+++ test/Tests/Application/ParametricData/ParametricDataErds_Test.cpp
@@ -72,7 +72,7 @@ TEST_GROUP(ParametricDataErds)
       appliancePersonality = PersonalityId_Default;
       DataModel_Write(dataModel, Erd_AppliancePersonality, &appliancePersonality);
 
-      for(int i = 0; i < PersonalityId_Max; i++)
+      for(uint8_t i = 0; i < PersonalityId_Max; i++)
       {
          personalites[i] = &personalitesData[i];
       }
